ability to remove operators, confirmations when removing groups/operators, capability checks
git-svn-id: https://webim.svn.sourceforge.net/svnroot/webim/trunk@525 c66351dc-e62f-0410-b875-e3a5c0b9693f

+++ src/messenger/webim/operator/groups.php
@@ -19,21 +19,30 @@ $operator = check_login();
 
 if( isset($_GET['act']) && $_GET['act'] == 'del' ) {
 	
-	// TODO check permissions
-	
-	$groupid = verifyparam( "gid", "/^(\d{1,9})?$/");
+	$groupid = isset($_GET['gid']) ? $_GET['gid'] : "";
 
-	$link = connect();
-	perform_query("delete from chatgroup where groupid = $groupid",$link);
-	perform_query("delete from chatgroupoperator where groupid = $groupid",$link);
-	perform_query("update chatthread set groupid = 0 where groupid = $groupid",$link);
-	mysql_close($link);
-	header("Location: $webimroot/operator/groups.php");
-	exit;
+	if( !preg_match( "/^\d+$/", $groupid )) {
+		$errors[] = "Cannot delete: wrong argument";
+	}
+	
+	if( !is_capable($can_administrate, $operator)) {
+		$errors[] = "You are not allowed to remove groups";
+	}
+	
+	if( count($errors) == 0 ) {
+		$link = connect();
+		perform_query("delete from chatgroup where groupid = $groupid",$link);
+		perform_query("delete from chatgroupoperator where groupid = $groupid",$link);
+		perform_query("update chatthread set groupid = 0 where groupid = $groupid",$link);
+		mysql_close($link);
+		header("Location: $webimroot/operator/groups.php");
+		exit;
+	}
 }
 
 $page = array();
 $page['groups'] = get_groups(true);
+$page['canmodify'] = is_capable($can_administrate, $operator);
 
 prepare_menu($operator);
 start_html_output();
